Fetch currency when the app first launch
Get the currency from api when the app first launch
After that, try to update in BroadReceiver when the network connection changes

+++ TravelCalculator/app/src/main/java/Countries/CountryPickerLietener.java
@@ -1,5 +1,5 @@
 package Countries;
 
 public interface CountryPickerLietener {
-    void onSelectCountry(int Id, String displayName, int countryNum, String countryCode, String currencyCode, float currency, String currencyName, String currencySymbol, int flagDrawableResID);
+    void onSelectCountry(int Id, String displayName, int countryNum, String countryCode, String currencyCode, double currency, String currencyName, String currencySymbol, int flagDrawableResID);
 }
