+++ TravelCalculator/app/src/main/java/com/lasalle2020android/travelcalculator/UserSettingActivity.java
@@ -78,7 +78,7 @@ public class UserSettingActivity extends AppCompatActivity implements CountryPic
     }
 
     @Override
-    public void onSelectCountry(int Id, String displayName, int countryNum, String countryCode, String currencyCode, float currency, String currencyName, String currencySymbol, int flagDrawableResID) {
+    public void onSelectCountry(int Id, String displayName, int countryNum, String countryCode, String currencyCode, double currency, String currencyName, String currencySymbol, int flagDrawableResID) {
         selectedCountry = new CountryModel();
         selectedCountry.setId(Id);
         selectedCountry.setDisplayName(displayName);
